Create sum of the square and cube by function.js
Make a function func that takes two numbers as parameters,
but return the sum of the square of the first number with the cube of
the second number.

For this, make a helper function square, which raises the number to
square, and the cube helper function to cube the number.